LayoutTests/imported/w3c/web-platform-tests/html/webappapis/animation-frames/callback-cross-realm-report-exception-expected.txt

CONSOLE MESSAGE: Error: PASS


PASS requestAnimationFrame() reports the exception from its callback in the callback's global object
